refactor(ghostmap): split observability decision-line/lifecycle formatters into GhostMapPresence.Observability.cs
